ReArtifyAI code/Re-Artify AI TEST/requirements.txt
diffusers==0.30.3
Flask==3.0.3
Flask_Cors==5.0.0
matplotlib==3.9.2
numpy==2.1.3
opencv_python==4.10.0.84
Pillow==11.0.0
torch==2.4.1
torchvision==0.19.1
